+++ test/game.test.js
@@ -1,10 +1,6 @@
-import game from './mock/game';
+import gameMock from './gameMock';
 
 test('Receive an object in return when call gameRun', () => {
-  const game = game.startGame();
+  const { game } = gameMock;
   expect(typeof game).toBe('object');
-});
-
-test('Expect to see the object that contains all the games scenes', () => {
-  expect(typeof game.scene.scenes).toBe('object');
 });
